requirements.txt
certifi==2026.4.22
cffi==2.0.0
click==8.3.3
colorama==0.4.6
coverage==7.13.5
cryptography==46.0.7
flake8==7.3.0
greenlet==3.3.2
iniconfig==2.3.0
markdown-it-py==4.0.0
mccabe==0.7.0
mdurl==0.1.2
packaging==26.0
pluggy==1.6.0
pycodestyle==2.14.0
pycparser==3.0
pyflakes==3.4.0
Pygments==2.20.0
PyJWT==2.12.1
pytest==9.0.3
pytest-cov==7.1.0
python-dotenv==1.2.2
rich==15.0.0
sentry-sdk==2.58.0
SQLAlchemy==2.0.48
typing_extensions==4.15.0
urllib3==2.6.3
